moving blocking call to a use case
